separate revalidate cli and updating all the readme files.
