+++ src/mcpeval/cli/mcp_task_revalidator/__init__.py
@@ -1,0 +1,7 @@
+"""
+MCP Task Revalidator CLI
+
+A command-line tool for revalidating task descriptions based on actual tool conversations.
+This tool takes verified tasks with conversation history and updates their descriptions
+to better reflect what actually happened during task execution.
+"""
