+++ src/mcpeval/cli/mcp_task_revalidator/__main__.py
@@ -1,0 +1,12 @@
+"""
+Main entry point for executing the mcp_task_revalidator module.
+
+This redirects to the main CLI.
+"""
+
+if __name__ == "__main__":
+    print("Please use the main CLI instead:")
+    print("python -m mcpeval.cli.main revalidate-tasks --help")
+    import sys
+
+    sys.exit(1)
